requirements.txt
# Core geospatial
numpy>=1.24
matplotlib>=3.7
rasterio>=1.3
shapely>=2.0
pyproj>=3.5
Pillow>=9.5

# Deep learning
torch>=2.0
torchvision>=0.15
segmentation-models-pytorch>=0.3.3

# App
streamlit>=1.28

# Optional (install if possible, not blocking)
# geopandas>=0.13
# pandas>=2.0
# scikit-image>=0.21
